requirements.txt
selenium
pytest
webdriver_manager
allure-pytest